Geração de relatórios e denúncias;

+++ projeto-pombo/src/main/java/pombo/spring/projeto_pombo/service/UsuarioService.java
@@ -8,6 +8,8 @@ import java.util.UUID;
 import javax.management.RuntimeErrorException;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Example;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 
 import pombo.spring.projeto_pombo.exception.ProjetoPomboException;
@@ -15,6 +17,7 @@ import pombo.spring.projeto_pombo.model.entity.Pruu;
 import pombo.spring.projeto_pombo.model.entity.Usuario;
 import pombo.spring.projeto_pombo.model.repository.PruuRepository;
 import pombo.spring.projeto_pombo.model.repository.UsuarioRepository;
+import pombo.spring.projeto_pombo.model.seletor.UsuarioSeletor;
 
 @Service
 public class UsuarioService {
@@ -25,7 +28,11 @@ public class UsuarioService {
 	@Autowired
 	private PruuRepository pruuRepository;
 	
-	public Usuario inserirUsuario(Usuario novoUsuario) {
+	@Autowired
+	private PruuService pruuService;
+	
+	public Usuario inserirUsuario(Usuario novoUsuario) throws ProjetoPomboException {
+		verificarUsuarioExistente(novoUsuario);
 		return usuarioRepository.save(novoUsuario);
 	}
 	
@@ -41,11 +48,16 @@ public class UsuarioService {
 		return usuarioRepository.save(usuarioAlterado);	
 	}
 	
-	public void deletarUsuarioPorId(String id) {
+	public void deletarUsuarioPorId(String id) throws ProjetoPomboException {
+		List<Pruu> pruusUsuario = pruuService.listarTodosPruusPorIdUsuario(id);
+		if(!pruusUsuario.isEmpty()) {
+			throw new ProjetoPomboException("Usuário com pruus criados não podem ser deletados!");
+		}
+		
 		usuarioRepository.deleteById(id);
 	}
 	
-	public void darLike(String idUsuario, String idPruu) {
+	public void darLike(String idUsuario, String idPruu) throws ProjetoPomboException {
 		Optional<Usuario> usuarioOpt = usuarioRepository.findById(idUsuario);
 		Optional<Pruu> pruuOpt = pruuRepository.findById(idPruu);
 		
@@ -63,7 +75,7 @@ public class UsuarioService {
 				usuarioRepository.save(usuario);
 			}
 		} else {
-			throw new RuntimeException("Usuario ou Pruu não encontrados.");
+			throw new ProjetoPomboException("Usuario ou Pruu não encontrados.");
 		}
 	} 
 	
@@ -89,5 +101,27 @@ public class UsuarioService {
 	    return false; 
 	}
 
+	public List<Usuario> pesquisarComSeletor(UsuarioSeletor seletor){
+		if(seletor.temPaginacao()) {
+			int numPage = seletor.getPagina();
+			int TamanhoPage = seletor.getLimite();
+			
+			PageRequest pagina = PageRequest.of(numPage - 1, TamanhoPage);
+			return usuarioRepository.findAll(pagina).toList();
+		}
+		
+		return usuarioRepository.findAll();
+	}
 	
+	public void verificarUsuarioExistente(Usuario usuarioVerificacao) throws ProjetoPomboException {
+		List<Usuario> usuarios = usuarioRepository.findAll();
+		
+		for (Usuario usuario : usuarios) {
+			if (usuario.getCpf().equals(usuarioVerificacao.getCpf())) {
+				throw new ProjetoPomboException("CPF já cadastrado no sistema!");
+			}else if(usuario.getEmail().equalsIgnoreCase(usuarioVerificacao.getEmail())){
+				throw new ProjetoPomboException("Email já cadastrado no sistema!");
+			}
+		}
+	}
 }
